+++ test/unit/auth-handlers.test.js
@@ -525,6 +525,10 @@ describe('Auth Handler:', function () {
                 'oauth_consumer_key', 'oauth_token', 'oauth_signature_method', 'oauth_timestamp', 'oauth_nonce',
                 'oauth_version', 'oauth_signature'
             ]);
+            // All the query paramters added by runtime must have `system: true` property
+            _.forEach(request.url.query.members, function (param) {
+                expect(param.system).to.be(true);
+            });
             expect(request.auth.parameters().toObject()).to
                 .eql({
                     consumerKey: 'RKCGzna7bv9YD57c',
@@ -591,7 +595,8 @@ describe('Auth Handler:', function () {
             expect(request.url.query.all().length).to.be(1);
             expect(request.url.query.all()[0]).to.eql({
                 key: 'access_token',
-                value: requestObj.auth.oauth2.accessToken
+                value: requestObj.auth.oauth2.accessToken,
+                system: true
             });
         });