+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Postman Runtime Changelog
 
 ### Unreleased
+* Runtime now adds `system: true` to all the query parameters that it sets
 
 ### v7.0.1 (November 8, 2017)
 * :bug: Fixed a bug where the assertions for legacy `tests` failures did not have an `error` object.